refactor(workspace): reorganize into filesystem/ and sandbox/ folders
- Move filesystem-related files to workspace/filesystem/
  - filesystem.ts, local-filesystem.ts, fs-utils.ts, file-read-tracker.ts
- Move sandbox-related files to workspace/sandbox/
  - sandbox.ts, local-sandbox.ts, native-sandbox/
- Add index.ts files for each folder
- Update all imports to use new paths
