+++ packages/core/src/workspace/filesystem/filesystem.ts
@@ -19,7 +19,7 @@
  * ```
  */
 
-import type { Lifecycle, ProviderStatus } from './lifecycle';
+import type { Lifecycle, ProviderStatus } from '../lifecycle';
 
 // =============================================================================
 // Core Types
