+++ packages/core/src/workspace/filesystem/index.ts
@@ -1,0 +1,4 @@
+export * from './filesystem';
+export * from './local-filesystem';
+export * from './fs-utils';
+export * from './file-read-tracker';
