+++ packages/core/src/workspace/sandbox/local-sandbox.ts
@@ -15,7 +15,7 @@ import * as os from 'node:os';
 import * as path from 'node:path';
 import { promisify } from 'node:util';
 
-import type { ProviderStatus } from './lifecycle';
+import type { ProviderStatus } from '../lifecycle';
 import type { IsolationBackend, NativeSandboxConfig } from './native-sandbox';
 import { detectIsolation, isIsolationAvailable, generateSeatbeltProfile, wrapCommand } from './native-sandbox';
 import type { WorkspaceSandbox, SandboxInfo, ExecuteCommandOptions, CommandResult } from './sandbox';
