+++ packages/core/src/workspace/sandbox/native-sandbox/detect.ts
