+++ packages/core/src/workspace/sandbox/native-sandbox/seatbelt.ts
