+++ packages/core/src/workspace/sandbox/native-sandbox/wrapper.ts
