+++ packages/core/src/workspace/tree-formatter.test.ts
@@ -4,7 +4,7 @@ import * as os from 'node:os';
 import * as path from 'node:path';
 import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 
-import { LocalFilesystem } from './local-filesystem';
+import { LocalFilesystem } from './filesystem';
 import { formatAsTree, formatEntriesAsTree } from './tree-formatter';
 
 describe('tree-formatter', () => {
